make addon api folder
